pyproject.toml
[project]
name = "calendar-bot"
version = "0.1.0"
description = "Телеграм бот для работы с календарем"
authors = [
    {name = "Ваше Имя", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.18.0",
    "sqlalchemy>=1.4.0,<2.0.0",
    "google-auth-oauthlib>=1.2.1",
    "google-api-python-client>=2.162.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.11.13",
    "pydantic>=2.10.6",
    "pytz>=2025.2",
    "black>=25.1.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "google_auth_oauthlib.*",
    "googleapiclient.*",
    "sqlalchemy.*",
    "google.*",
    "database"
]
ignore_missing_imports = true
disallow_untyped_defs = false
